CheckBox: extract CheckBoxIcon component (#18844)
