+++ js/ui/themes.d.ts
@@ -44,3 +44,4 @@ export default class themes {
 
 export function current(): string;
 export function isMaterial(theme: string): boolean;
+export function isCompact(theme: string): boolean;
