+++ js/ui/themes.js
@@ -336,6 +336,10 @@ export function isDark(themeName) {
     return isTheme('dark', themeName);
 }
 
+export function isCompact(themeName) {
+    return isTheme('compact', themeName);
+}
+
 export function isWebFontLoaded(text, fontWeight) {
     const testedFont = 'Roboto, RobotoFallback, Arial';
     const etalonFont = 'Arial';
@@ -443,6 +447,7 @@ export default {
     ready: themeReady,
     waitWebFont,
     isWebFontLoaded,
+    isCompact,
     isDark,
     isGeneric,
     isMaterial,
